Update the highlight area for SNR calculation
The 'stacked_periodogram' function was storing only the last 'best_P' in the 'results' dictionary. Therefore, 'plot_SNR' option from 'plot_stacked_periodogram_heatmap' function was only considering the last 'best_P', plotting a vertical region around that x-value in the stack plot and implying that the SNR was being calculated from the values within that highlighted region. This was not the case. For well-behaved and strong signals, this change makes little difference, but in most cases of interest, it is very important to see what is really happening.
This patch addresses this issue and now the code shows the real area being considered when calculating the SNR plot.